converted python lists to numpy arrays.
Due to the fact that numpy arrays have a more complex data structure
than python lists, this has caused the program to run slower for smaller
lists. However the use of numpy lists will be useful when calculating
large arrays as will hopefully be done in the final run of the project.

The old non-numpy file is found in discfunc.old